fix bug with cast Long to Integer

+++ src/main/java/com/car/dealer/dao/ManagerDaoImpl.java
@@ -45,7 +45,7 @@ public class ManagerDaoImpl implements ManagerDao {
         query.setParameter("manager", manager);
 
         Brand brand = (Brand) ((Object[])query.uniqueResult())[0];
-        Integer count = (Integer) ((Object[])query.uniqueResult())[1];
+        Long count = (Long) ((Object[])query.uniqueResult())[1];
 
         return new BrandMaxCount(brand, count);
     }
